+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 3.7-beta
 - RTL layout issues #214 in rtl.css
 - INSERT INTO case to Replace empty strings ('') with NULL values in database.inc.php
 - Add DISPLAY_NAME to CONFIG table in rosariosis.sql & Update.fnc.php
+- Fix #221 Delete Addresses & Contacts info does not delete student in Student.php & Address.inc.php
 
 Changes in 3.6.1
 ----------------